Remove `ResultFormatter` class and make static methods functions
